pyproject.toml
[project]
name = "manga-translator"
version = "1.0.0"
description = "Manga Translator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ultralytics>=8.0.0",
    "easyocr>=1.7.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "transformers==4.50",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "streamlit>=1.28.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "argostranslate>=1.9.1",
    "langdetect>=1.0.9",
    "scikit-image>=0.22.0",
    "scikit-learn>=1.3.0",
    "pydantic-settings",
    "loguru>=0.7.3",
    "python-doctr>=1.0.0",
    "manga-ocr>=0.1.14",
    "googletrans>=4.0.2",
    "paddleocr>=3.3.3",
    "paddlepaddle==3.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
]

[tool.ruff]
line-length = 130
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I", "E", "F", "W", "Q", "W291", "W293"]

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.isort]
case-sensitive = true
force-wrap-aliases = true
combine-as-imports = true
lines-after-imports = 2
split-on-trailing-comma = true
